requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
opencv-python==4.8.1.78
